+++ Source/Modules/ISHServer/ISHServer.15.psm1
@@ -1,5 +1,5 @@
 <#
-# Copyright (c) 2023 All Rights Reserved by the RWS Group.
+# Copyright (c) 2024 All Rights Reserved by the RWS Group.
 # 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ Set-Variable -Name "ISHServer:NETFramework" -Value "ndp48-x86-x64-allos-enu.exe"
 Set-Variable -Name "ISHServer:VisualBasicRuntime" -Value "vbrun60sp6.exe" -Scope "Script" -Option Constant
 Set-Variable -Name "ISHServer:MSOLEDBSQLRequiredVersion" -Value "18.2.1.0" -Scope "Script" -Option Constant
 Set-Variable -Name "ISHServer:MSOLEDBSQL" -Value "msoledbsql_18.3.0.0_x64.msi" -Scope "Script" -Option Constant
-Set-Variable -Name "ISHServer:DotNetHosting" -Value "dotnet-hosting-8.0.0-win.exe"  -Scope "Script" -Option Constant
+Set-Variable -Name "ISHServer:DotNetHosting" -Value "dotnet-hosting-6.0.1-win.exe"  -Scope "Script" -Option Constant
 
 $exportNames=@(
     #region Helpers